// schema.d.ts
export interface paths {
// ...
    "/v2.0/{season}/events": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Event Listings
         * @description The event listings API returns all FTC official regional events in a particular season. You can specify an `eventCode` if you would only like data about one specific event. If you specify an `eventCode` you cannot specify any other optional parameters. Alternately, you can specify a `teamNumber` to retrieve only the listings of events being attended by the particular team. If you specify a `teamNumber` you cannot specify an `eventCode`.
         *
         *     The response for event listings contains a special field called divisionCode. For example, the FIRST Championship contains two Divisions. As an example of a response, the event listings for a Division will have a divisionCode that matches the FIRST Championship event code (as they are divisions of that event). This allows you to see the full structure of events, and how they relate to each other.
         */
        get: {
            parameters: {
                query?: {
                    /** @description Case insensitive alphanumeric `eventCode` of the event about which details are requested. */
                    eventCode?: string;
                    /** @description Numeric `teamNumber` of the team from which the attending event listings are requested. */
                    teamNumber?: number;
                };
                header?: never;
                path: {
                    /** @description Numeric year from which the event listings are requested. Must be 4 digits */
                    season: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"][
                            "SeasonEventListingsModel_Version2"
                        ];
                    };
                };
                /** @description Unauthorized */
                401: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v2.0/{season}/matches/{eventCode}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Event Match Results
         * @description The match results API returns the match results for all matches of a particular event in a particular season. Match results are only available once a match has been played, retrieving info about future matches requires the event schedule API. You cannot receive data about a match that is in progress. You can, however, request the Hybrid Schedule if you would like data about upcoming and played matches at the same time.
         *
         *     If you specify the `matchNumber`, `start` and/or `end` optional parameters, you must also specify a `tournamentLevel`. If you specify the `teamNumber` parameter, you cannot specify a `matchNumber` parameter. If you specify the `matchNumber`, you cannot define a start or end.
         *
         *     Note: If you specify `start`, and it is higher than the maximum match number at the event, you will not receive any match results in the response. The same is true in reverse for the `end` parameter.
         */
        get: {
            parameters: {
                query?: {
                    /** @description Required tournamentLevel of desired score details. */
                    tournamentLevel?: "qual" | "playoff";
                    /** @description `teamNumber` to search for within the results. Only returns match results in which the requested team was a participant. */
                    teamNumber?: number;
                    /** @description specific single `matchNumber` of result. */
                    matchNumber?: number;
                    /** @description `start` match number for subset of results to return. */
                    start?: number;
                    /** @description `end` match number for subset of results to return (inclusive). */
                    end?: number;
                };
                header?: never;
                path: {
                    /** @description Numeric year of the event from which the match results are requested. Must be 4 digits. */
                    season: number;
                    /** @description Case insensitive alphanumeric `eventCode` of the event from which the results are requested. Must be at least 3 characters. */
                    eventCode: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"][
                            "EventMatchResultsModel_Version2"
                        ];
                    };
                };
                /** @description Unauthorized */
                401: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v2.0/{season}/scores/{eventCode}/{tournamentLevel}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Score Details
         * @description The score details API returns the score detail for all matches of a particular event in a particular season and a particular tournament level. Score details are only available once a match has been played, retrieving info about future matches requires the event schedule API. You cannot receive data about a match that is in progress.
         */
        get: {
            parameters: {
                query?: {
                    /** @description `teamNumber` to search for within the results. Only returns details in which the requested team was a participant. */
                    teamNumber?: number;
                    /** @description specific single `matchNumber` of result. */
                    matchNumber?: number;
                    /** @description `start` match number for subset of results to return (inclusive). */
                    start?: number;
                    /** @description `end` match number for subset of results to return (inclusive). */
                    end?: number;
                };
                header?: never;
                path: {
                    /** @description Numeric year of the event from which the match results are requested. Must be 4 digits. */
                    season: number;
                    /** @description Case insensitive alphanumeric eventCode of the event from which the details are requested. Must be at least 3 characters. */
                    eventCode: string;
                    /** @description Required tournamentLevel of desired score details. */
                    tournamentLevel: "qual" | "playoff";
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Success */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json":
                            components["schemas"]["MatchScoresModel"];
                    };
                };
                /** @description Unauthorized */
                401: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        ScoreDetailModel_2023: {
            matchLevel?: components["schemas"]["FTCEventLevel"];
            /** Format: int32 */
            matchSeries?: number;
            /** Format: int32 */
            matchNumber?: number;
            /** Format: int32 */
            randomization?: number;
            alliances?:
                | components["schemas"]["ScoreDetailAllianceModel_2023"][]
                | null;
        };
        /** IntoTheDeep Alliance Score Details */
        ScoreDetailModel_2024: {
            matchLevel?: components["schemas"]["FTCEventLevel"];
            /** Format: int32 */
            matchSeries?: number;
            /** Format: int32 */
            matchNumber?: number;
            /** Format: int32 */
            randomization?: number;
            alliances?:
                | components["schemas"]["ScoreDetailAllianceModel_2024"][]
                | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}